Update ARCANA... (still needs to be refactored)
